add CategoryGateway
