+++ service_restaurant/src/main/kotlin/org/thechance/service_restaurant/data/gateway/RestaurantGatewayImp.kt
@@ -18,80 +18,8 @@ import org.thechance.service_restaurant.entity.Restaurant
 
 @Single
 class RestaurantGatewayImp(private val container: DataBaseContainer) : RestaurantGateway {
-
     private val restaurantCollection by lazy { container.database.getCollection<RestaurantCollection>() }
-
     private val categoryCollection by lazy { container.database.getCollection<CategoryCollection>() }
-
-    //region Category
-    override suspend fun getCategories(page: Int, limit: Int): List<Category> {
-        return categoryCollection.find(CategoryCollection::isDeleted eq false)
-            .paginate(page, limit).toList().toEntity()
-    }
-
-    override suspend fun getCategory(categoryId: String): Category? {
-        return categoryCollection.aggregate<CategoryCollection>(
-            match(and(CategoryCollection::id eq ObjectId(categoryId), CategoryCollection::isDeleted eq false)),
-            project(CategoryCollection::name, CategoryCollection::id)
-        ).toList().firstOrNull()?.toEntity()
-    }
-
-    override suspend fun getRestaurantsInCategory(categoryId: String): List<Restaurant> {
-        return categoryCollection.aggregate<CategoryRestaurantCollection>(
-            match(CategoryCollection::id eq ObjectId(categoryId)),
-            lookup(
-                from = "restaurantCollection",
-                resultProperty = CategoryRestaurantCollection::restaurants,
-                pipeline = listOf(match(RestaurantCollection::isDeleted eq false)).toTypedArray()
-            ),
-        ).toList().first().restaurants.toEntity()
-    }
-
-    override suspend fun addCategory(category: Category): Boolean {
-        return categoryCollection.insertOne(category.toCollection()).wasAcknowledged()
-    }
-
-    override suspend fun addRestaurantsToCategory(categoryId: String, restaurantIds: List<String>): Boolean {
-        val resultAddToRestaurant = restaurantCollection.updateMany(
-            RestaurantCollection::id `in` restaurantIds.toObjectIds(),
-            addToSet(RestaurantCollection::categoryIds, ObjectId(categoryId))
-        ).isSuccessfullyUpdated()
-
-        val resultAddToCategory = categoryCollection.updateOneById(
-            ObjectId(categoryId),
-            update = Updates.addEachToSet(CategoryCollection::restaurantIds.name, restaurantIds.toObjectIds())
-        ).isSuccessfullyUpdated()
-        return resultAddToCategory and resultAddToRestaurant
-    }
-
-    override suspend fun updateCategory(category: Category): Boolean {
-        return categoryCollection.updateOneById(
-            id = ObjectId(category.id),
-            update = category.toCollection(),
-            updateOnlyNotNullProperties = true
-        ).isSuccessfullyUpdated()
-    }
-
-    override suspend fun deleteCategory(categoryId: String): Boolean {
-        return categoryCollection.updateOneById(
-            id = ObjectId(categoryId),
-            update = Updates.set(CategoryCollection::isDeleted.name, true),
-        ).isSuccessfullyUpdated()
-    }
-
-    override suspend fun deleteRestaurantsInCategory(categoryId: String, restaurantIds: List<String>): Boolean {
-        val resultDeleteFromRestaurant = restaurantCollection.updateMany(
-            RestaurantCollection::id `in` restaurantIds.toObjectIds(),
-            pull(RestaurantCollection::categoryIds, ObjectId(categoryId))
-        ).isSuccessfullyUpdated()
-
-        val resultDeleteFromCategory = categoryCollection.updateOneById(
-            ObjectId(categoryId),
-            pullAll(CategoryCollection::restaurantIds, restaurantIds.toObjectIds())
-        ).isSuccessfullyUpdated()
-        return resultDeleteFromRestaurant and resultDeleteFromCategory
-    }
-    //endregion
 
     //region Restaurant
     override suspend fun getRestaurants(page: Int, limit: Int): List<Restaurant> {
